+++ changelog.md
@@ -12,6 +12,7 @@ Features:
 * Added `.load` command for loading extension libraries. (Thanks: [Zhiming Wang])
 * Add support for using `?` as a placeholder in the favorite queries. (Thanks: [Amjith])
 * Added shift-tab to select the previous entry in the completion menu. [Amjith]
+* Added `describe` and `desc` keywords.
 
 1.0.0
 =====